% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_flock_description}
\alias{clean_flock_description}
\title{Clean Flock Description data.}
\usage{
clean_flock_description(df, person)
}
\arguments{
\item{df}{data.frame Data frame to be cleaned.}

\item{person}{str Person data petains to, should be one of 'LN', 'MJ', 'ND', 'SB'.}
}
\value{
Dataframe of flock description with columns renamed.
}
\description{
Mostly this is just a case of renaming variables. THe data points as captured are cleaner.
}
