% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_gps}
\alias{clean_gps}
\title{Clean a GPS file.}
\usage{
clean_gps(
  file,
  format = "csv",
  skip = 42,
  col_select = c("name", "lat", "lon", "ele", "time"),
  drop_two = TRUE
)
}
\arguments{
\item{file}{str Path to file to be loaded.}

\item{format}{str Format of GPS file, currently on "csv" is supported.}

\item{skip}{int Number of rows in header of ".csv" to skip. Default is 42.}

\item{col_select}{list List of columns to select. Default is c("name", "lat", "lon", "ele", "time")}

\item{drop_two}{bool Whether to drop the first two rows of GPS data points. Default is TRUE.}
}
\value{
Dataframe of GPS positions loaded from '.csv' file.
}
\description{
Extracts GPS data from the supplied file. The 'person' variable is derived from patterns in the file name, currently
'LN', 'MJ', 'ND', 'SB' and 'SJB'. Typically when the GPS data is exported to CSV the files have 42 rows in their
header which are not required, hence the 'skip' value being 42. Many GPS tracks appear to have a slight lag in
obtaining position and the first two recorded points retain the lat/lon/elevation from the last recorded trip. These
can optionally be removed using the 'drop_two' boolean option which defaults to TRUE.
}
