% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_empty_dataframes}
\alias{create_empty_dataframes}
\title{Setup empty dataframes for the different components.}
\usage{
create_empty_dataframes()
}
\value{
List of empty dataframes used in the Shiny application.
}
\description{
Setup empty dataframes for the different components.
}
