% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataTableServer}
\alias{dataTableServer}
\title{DataTable module server function}
\usage{
dataTableServer(id, reactive_data, empty_df)
}
\arguments{
\item{id}{str ID for namespace within which UI element(s) live.}

\item{reactive_data}{function Data for rendering as DataTable, expected to be a reactive function.}

\item{empty_df}{dataframe Empty dataframe indicating data types.}
}
\value{
\code{NULL}
}
\description{
DataTable module server function
}
