% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataTableUI}
\alias{dataTableUI}
\title{DataTable module UI function}
\usage{
dataTableUI(id)
}
\arguments{
\item{id}{str ID for namespace within which UI element(s) will live.}
}
\value{
DataTable UI output.
}
\description{
DataTable module UI function
}
