% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_table_debug}
\alias{db_table_debug}
\title{Query a database table to check contents for debugging.}
\usage{
db_table_debug(conn, table)
}
\arguments{
\item{conn}{Database connection.}

\item{table}{str Name of table to query.}
}
\value{
Prints the table.
}
\description{
Query a database table to check contents for debugging.
}
