% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deduplicate_flock}
\alias{deduplicate_flock}
\title{De-duplicate ringed entries from flock description.}
\usage{
deduplicate_flock(df)
}
\arguments{
\item{df}{dataframe Dataframe for de-duplicating.}
}
\value{
Dataframe with ringed birds deduplicated.
}
\description{
When entering data there should be no duplicated individuals in the flock with rings. Un-ringed birds are
permissible, by virtue of sharing the characteristic of not having rings.
}
