% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_all_tables}
\alias{extract_all_tables}
\title{Extract all specified tables.}
\usage{
extract_all_tables(
  con,
  path,
  tables = c("Conditions", "Composition", "Description", "Interactions", "GPS")
)
}
\arguments{
\item{con}{Database connection.}

\item{path}{Path to export files to.}

\item{tables}{list List of tables as strings to be extracted.}
}
\value{
A list of extracted/exported files.
}
\description{
Extract all specified tables.
}
