% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_and_compress_data}
\alias{extract_and_compress_data}
\title{Extract and compress data from database to zip file.}
\usage{
extract_and_compress_data(zip_file, conn, input)
}
\arguments{
\item{zip_file}{str Zip filename.}

\item{conn}{Database connection}

\item{input}{List of tables to extract}
}
\value{
Returns nothing, exports database tables to CSV and compresses them.
}
\description{
Extract and compress data from database to zip file.
}
