% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_rings}
\alias{extract_rings}
\title{Extract rings}
\usage{
extract_rings(code, valid_codes, known_rings)
}
\arguments{
\item{code}{str The code from which rings and leg are to be extracted.}

\item{valid_codes}{list List of valid codes, if \code{code} is not in \code{valid_codes} it is not extracted.}

\item{known_rings}{list List of rings used.}
}
\value{
List of ring properties.\tabular{lll}{
   Name \tab Type \tab Description \cr
   \code{code} \tab str \tab The original full ring code. \cr
   \code{leg} \tab str \tab The leg denoted by the full rung code (the last letter). \cr
   \code{pit} \tab boolean \tab Whether a "pit" ring is indicated. \cr
   \code{pit_pos} \tab str \tab The position of the "pit" ring. \cr
   \code{bto} \tab str \tab The leg the BTO ring is on (inferred). \cr
   \code{first} \tab str \tab The first ring from the full ring code. \cr
   \code{second} \tab str \tab The second ring from the full ring code. \cr
}
}
\description{
See the document
\href{https://docs.google.com/document/d/10JnL50Fm5DneEl9PgqYLbxUDLo6wcnaUbGvY6e3j8Oc/edit?tab=t.0}{Reading Colour
Rings} for further details on the rules used in the coding system.
}
