% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_table_to_csv}
\alias{extract_table_to_csv}
\title{Extract tables from a database and write to '.csv'.}
\usage{
extract_table_to_csv(table, con, path)
}
\arguments{
\item{table}{str Table to extract data from. All columns will be exported.}

\item{con}{Database connection.}

\item{path}{Path to export to.}
}
\value{
N/A nothing is returned, table is written to \verb{/path/table.csv}.
}
\description{
Extract tables from a database and write to '.csv'.
}
