% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_inputs}
\alias{filter_inputs}
\title{Filter a list of inputs for a subset and remove those that are to be excluded.}
\usage{
filter_inputs(inputs, filter, exclude)
}
\arguments{
\item{inputs}{list A list of strings which are to be filtered.}

\item{filter}{str The string on which to filter, should be one of \verb{^composition_}, \verb{^description_}, and
\verb{^interactions_}.}

\item{exclude}{list A list of inputs to exclude.}
}
\value{
List of input fields filtered for the subset specified but with some fields excluded.
}
\description{
We wish to reset input fields, to do so we need a list of labels used for \code{input} objects. This is provided by the
\code{all_inputs()} function which returns \emph{all} inputs for a given session. We have three areas we wish to subset
\code{composition}, \code{description}, and \code{interactions} but there are a small number of each that we do \emph{not} want to
reset because they are either automatically incremented or take their values from other areas that have been input
(e.g. when entering in \code{composition_} fields we wish to retain the \code{flock_number} that has been entered from the
\code{flock} description; for interactions we wish the list of possible flocks under \code{flock_a} and \code{flock_b} to be
based on the flocks that have been entered already.
}
