% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_none_column}
\alias{remove_none_column}
\title{Remove 'none' column from a dataframe if it exists.}
\usage{
remove_none_column(df)
}
\arguments{
\item{df}{data.frame Data frame to be augmented.}
}
\value{
Dataframe with column called \code{none} removed.
}
\description{
Remove 'none' column from a dataframe if it exists.
}
