% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_dt}
\alias{render_dt}
\title{Render a reactive dataframe as an editable DataTable.}
\usage{
render_dt(
  reactiveData,
  editable = "cell",
  exclude_cols = NULL,
  server = TRUE,
  rownames = FALSE,
  options = list(),
  ...
)
}
\arguments{
\item{reactiveData}{function Data for rendering as table, expected to be a reactive function.}

\item{editable}{str|list Controls table elements to be editable (\code{cell|row|column|all} or \code{list()}, default \code{cell}).}

\item{exclude_cols}{int Vector of zero-indexed column indices for which editing should be disabled (default \code{NULL}).}

\item{server}{bool Whether to run server side or not (default \code{TRUE}).}

\item{rownames}{bool Whether to display row names (default = \code{FALSE}).}

\item{options}{list List of DataTable options (default = \code{list()}).}

\item{...}{Other parameters to be passed on to \code{DT::renderDT()}.}
}
\value{
Rendered DataTable.
}
\description{
Render a reactive dataframe as an editable DataTable.
}
