% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_table}
\alias{render_table}
\title{Render a dataframe table as a Shiny.}
\usage{
render_table(df, striped = TRUE)
}
\arguments{
\item{df}{dataframe Dataframe for rendering as table.}

\item{striped}{bool Whether the table should be striped or not (default \code{TRUE}).}
}
\value{
Shiny rendered table.
}
\description{
Render a dataframe table as a Shiny.
}
