% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_data}
\alias{save_data}
\title{Save data to SQLite database}
\usage{
save_data(data, db_path = db_path, table, append = TRUE, overwrite = FALSE)
}
\arguments{
\item{data}{dataframe Dataframe of data to be added to a table in the database.}

\item{db_path}{str Path to database.}

\item{table}{str Table to add data to.}

\item{append}{bool Whether to append the data.}
}
\value{
Nothing returned.
}
\description{
This generic function can be used to populate lookup tables when testing the database in memory or more generally
when running the database to add data that has been entered.
}
