% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_columns}
\alias{tidy_columns}
\title{Tidy other bird species columns.}
\usage{
tidy_columns(df, expected_cols)
}
\arguments{
\item{df}{data.frame Data frame to be augmented.}

\item{expected_cols}{list\link{str} List of columns that the data frame should hold.}
}
\value{
Dataframe with missing columns added and \code{none} column removed.
}
\description{
Ensures there is a boolean column for each possible "Other species" that might be observed. This is required because
if a species is \emph{not} observed then after reshaping to wide there is no boolean column for that species. As a
consequence combining the dataframe with other flock description observations during that session will fail and
attempting to update the database will also fail.
}
\details{
Further when recording "Other Species" the default check box that is selected is "None" so that we do not observe an
error when reshaping the data if no "Other Species" are observed. As a consequence it is possible for the reshaped
data frame to contain a column called \code{none} which is irrelevant (whether this is because it was not unchecked when
adding "Other Species" or there were genuinely no "Other Species" observed). A call is therefore made to the
\code{remove_none_column()} function so that if the "None" check box was checked the resulting column is removed.
}
