% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_all_rings}
\alias{update_all_rings}
\title{Update individual ring options based on global selection.}
\usage{
update_all_rings(rings, session)
}
\arguments{
\item{rings}{list A named list of ring attributes, returned by \code{extract_ring()}.}

\item{session}{A shiny session which is to be updated.}
}
\value{
List of ring attributes. \code{pit} (boolean) \code{pit_pos} (str), \code{leg} (str), \code{code} (str), \code{bto} (str).
}
\description{
Takes the selected rings that have been split using \code{extract_ring()} and updates the displayed ring for individual
left/right top/left options.
}
\details{
The logic handles PIT rings first, then BTO rings before handling any that are just coloured rings alone. BTO rings
are always present but are rarely explicitly stated.
}
