% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_certainty}
\alias{update_certainty}
\title{Update individual ring certainty based on overall certainty}
\usage{
update_certainty(ring_certainty, tag, session)
}
\arguments{
\item{ring_certainty}{bool Whether the overall certainty is 'TRUE' or 'FALSE'.}

\item{tag}{str The individual certainty box to be updated ('left_top' | 'left_bottom' | 'right_top' | 'right_bottom')}

\item{session}{The Shiny session to update.}
}
\value{
Nothing, updates ring_certainty fields input for given tag and session.
}
\description{
If the "Certainty" box is ticked for the overall ring code this function updates the relevant individual "Certainty"
checkbox for the given "tag"
}
