% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_date}
\alias{update_date}
\title{Update a \code{shiny::dateInput()} field.}
\usage{
update_date(date, tag, session)
}
\arguments{
\item{date}{Date/time to update with.}

\item{tag}{str The \code{inputID} of the element to be updated.}

\item{session}{Shiny session to update from.}
}
\value{
Nothing, updates date input for given session.
}
\description{
Update a \code{shiny::dateInput()} field.
}
