% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_ring}
\alias{update_ring}
\title{Update the selected value of a specific leg position.}
\usage{
update_ring(session, tag, selected)
}
\arguments{
\item{session}{Shiny session which is to be updated.}

\item{tag}{str A string indicating the leg/position to be updated (one of \code{left_top}, \code{left_bottom}, \code{right_top},
\code{right_bottom}).}

\item{selected}{str The value to update the given tag with.}
}
\value{
Nothing, updates a ring field input for given session.
}
\description{
Update the selected value of a specific leg position.
}
