% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_rings_when_not_ringed}
\alias{update_rings_when_not_ringed}
\title{Update ring component fields.}
\usage{
update_rings_when_not_ringed(session)
}
\arguments{
\item{session}{The session to be updated.}
}
\value{
Updates input fields for rings in the supplied session.
}
\description{
This function is called when the \code{composition_ringed} field is changed to "No" (\code{FALSE}) and clears all ring fields
so that no erroneous data is submitted.
}
