% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_time}
\alias{update_time}
\title{Update a \code{shinyTime::timeInput()} field.}
\usage{
update_time(date, tag, session)
}
\arguments{
\item{date}{Date/time to update with.}

\item{tag}{str The \code{inputID} of the element to be updated.}

\item{session}{Shiny session to update from.}
}
\value{
Nothing, updates time input for given session.
}
\description{
Update a \code{shinyTime::timeInput()} field.
}
