% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_rings}
\alias{extract_rings}
\title{Extract rings}
\usage{
extract_rings(code, valid_ring_combinations, valid_rings)
}
\arguments{
\item{code}{str The code from which rings and leg are to be extracted.}

\item{valid_ring_combinations}{list A list of valid ring combinations of the form "[ring1][ring2][leg]", e.g. \code{Y*WL}, \code{SdnrL},
\code{ryFR}, \code{BDL}. These will typically be taken from the \code{R/lookups.R} file which hard codes the available rings based
on an extract from an external database. New codes should be added and/or the list updated from a database extract
but users should be aware of possible errors (one identified and corrected is the presence of \code{UL} which should be
\code{UUL} and has been manually corrected).}

\item{valid_rings}{list List of individual rings used in ringing birds, combinations thereof form \code{code} and are
typically in \code{valid_ring_combinations}, however because of uncertainty when capturing data in the field it is
possible that a \code{code} formed from \code{valid_rings} may not exist in \code{valid_ring_combinations}.}
}
\value{
List of ring properties.\tabular{lll}{
   Name \tab Type \tab Description \cr
   \code{code} \tab str \tab The original full ring code. \cr
   \code{leg} \tab str \tab The leg denoted by the full rung code (the last letter). \cr
   \code{pit} \tab boolean \tab Whether a "pit" ring is indicated. \cr
   \code{pit_pos} \tab str \tab The position of the "pit" ring. \cr
   \code{bto} \tab str \tab The leg the BTO ring is on (inferred). \cr
   \code{first} \tab str \tab The first ring from the full ring code. \cr
   \code{second} \tab str \tab The second ring from the full ring code. \cr
}
}
\description{
See the document
\href{https://docs.google.com/document/d/10JnL50Fm5DneEl9PgqYLbxUDLo6wcnaUbGvY6e3j8Oc/edit?tab=t.0}{Reading Colour
Rings} for further details on the rules used in the coding system.
}
